+++ core/data.js
@@ -4,7 +4,7 @@ export const data = {
   basedir: path.join('.', '.zap'),
   branch: path.join('.', '.zap', 'branch'),
   defaultBranch: 'main',
-  version: '2.0.0',
+  version: '3.0.1',
 };
 
 export default data;